This version attempts to get 'least common' info
When there are multiple returns from the Google location API, this checks for any 'common' fields across all those returned and uses that "smallest common set" as if it were unique.  I'm not sure exactly where it will help with this "2 lookups only" version, but it doesn't seem to hurt.